backend/foodgram/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
django-filter==22.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
Pillow==10.0.1
psycopg2-binary==2.9.3
gunicorn==20.1.0